requirements.txt
flask
flask-cors
numpy
pandas
sklearn
gensim
fasttext
natto-py
mecab-python3